requirements.txt
-e git+https://github.com/101dotxyz/GPTeam.git@a216364ea2dbb208e6501468c5e5b791594cc3d5#egg=101_ai
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.19.0
anthropic==0.2.10
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
backports.tarfile==1.2.0
black==23.12.1
blinker==1.5
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.2.1
click==8.1.7
click-log==0.4.0
colorama==0.4.6
colorlog==6.9.0
crashtest==0.4.1
dataclasses-json==0.5.14
deprecation==2.1.0
distlib==0.3.9
docutils==0.21.2
dotty-dict==1.3.1
dulwich==0.21.7
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
gotrue==1.3.1
h11==0.14.0
h2==4.1.0
hikari==2.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.26.5
Hypercorn==0.17.3
hyperdb-python==0.1.4
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
keyring==25.5.0
langchain==0.0.144
MarkupSafe==3.0.2
marshmallow==3.23.1
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nh3==0.2.19
numexpr==2.10.2
numpy==1.26.4
openai==0.27.10
openapi-schema-pydantic==1.2.4
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
postgrest==0.10.6
postgrest-py==0.10.6
priority==2.0.0
propcache==0.2.1
ptyprocess==0.7.0
pycparser==2.22
pydantic==1.10.19
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2023.4
PyYAML==6.0.2
quart==0.18.4
readme_renderer==44.0
realtime==1.0.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.54
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
taskgroup==0.0.0a4
tenacity==8.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2024.10.21.16
twine==3.8.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
utils==1.0.2
virtualenv==20.27.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.8
wolframalpha==5.1.3
wsproto==1.2.0
xattr==1.1.0
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
